+++ src/cryojax/simulator/_scattering_theory/multislice_scattering_theory.py
@@ -6,8 +6,8 @@ from jaxtyping import Array, Complex, Float, PRNGKeyArray
 from ...jax_util import error_if_not_fractional
 from .._image_config import AbstractImageConfig
 from .._transfer_theory import WaveTransferTheory
+from .._volume import AbstractVolumeRepresentation
 from .._volume_integrator import AbstractMultisliceIntegrator
-from .._volume_parametrisation import AbstractVolumeRepresentation
 from .base_scattering_theory import AbstractWaveScatteringTheory
