+++ src/cryojax/simulator/_scattering_theory/weak_phase_scattering_theory.py
@@ -6,8 +6,8 @@ from jaxtyping import Array, Complex, Float, PRNGKeyArray
 from .._image_config import AbstractImageConfig
 from .._solvent_2d import AbstractRandomSolvent2D
 from .._transfer_theory import ContrastTransferTheory
+from .._volume import AbstractVolumeRepresentation
 from .._volume_integrator import AbstractDirectIntegrator, AbstractDirectVoxelIntegrator
-from .._volume_parametrisation import AbstractVolumeRepresentation
 from .base_scattering_theory import AbstractWeakPhaseScatteringTheory
