+++ src/cryojax/simulator/_volume_integrator/multislice_integrator/fft_multislice_integrator.py
@@ -6,7 +6,7 @@ from jaxtyping import Array, Complex, Float
 
 from ....ndimage import fftn, ifftn, map_coordinates
 from ..._image_config import AbstractImageConfig
-from ..._volume_parametrisation import RealVoxelGridVolume
+from ..._volume import RealVoxelGridVolume
 from .base_multislice_integrator import AbstractMultisliceIntegrator
